moved videoEntry to validators
