+++ apps/web/src/lib/validators/videoEntry.ts
@@ -1,7 +1,7 @@
 import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import type { z } from "zod";
 
-import { videoEntries } from "./schema";
+import { videoEntries } from "~/server/db/schema";
 
 export const videoEntrySchema = createSelectSchema(videoEntries);
 export type VideoEntry = z.infer<typeof videoEntrySchema>;
